ROUNDS are now 1 less.
For performing the shifts in the round constants, we need to do a
rotate of the vector of round constants to the left. When writing
generic code, we had to provide a proof that this vector is indeed a
vector of length (S n) for some n. Ofcourse, rounds are much more than
0 but this is not something Coq is able to see easily.

Instead of going through some complicated proofs here, we just follow
the convention that the rounds parameter is one less than the total
rounds. That way the round constant vector is defined to be a vector
of (S ROUNDS) many value. This allows us to preform the rotates
without any difficulty.