rearranged arguments to help instance search
The order in which the arguments were given to the functions
binOpUpdate and uniOpUpdate resulted in type resolution not working
when using instructions of the kind x ::= Ox "abcd". The reason for
this is that the expression like Ox "abcd" needs to know what type is
expected in the context where it appears to determine its type.  If
the type expected is Word16, the expression will resolve well. However
if the context demands an expression of type Word64 then Ox "abcd"
will give a type error.

The old order of arguments for binOpUpdate and uniOpUpdate meant that
it will first try to resolve all the class variables with no clue on
type ty to begin with. With the current argument order, resolution of
the type ty will happen as soon as we have provided the lhs
argument. This leads to successful resolution of the class for the rhs
expression and hence rhs like Ox "abcd" works as expected (it now
knows what type is expected).