remove dependency on r < ROUNDS proof.
Due to the difference in the implementation of the tactics across
versions the type of the resulting bounds function was changing.  The
verse tactic used to use the omega tatics that required the proof that
r < ROUNDS and hence the BLAKE_ROUND function had a dependency on the
parameter. However, with the change to lia this dependency went of
resulting in the types mismatching across version.